requirements.txt
django==4.0.6
djangorestframework==3.13.1
djangorestframework-simplejwt=5.2.0
python-dotenv==0.20.0
psycopg2-binary==2.8.6
django-cors-headers==3.13.0